[refactoring] Properly name plugin
