+++ api/lib/infrastructure/plugins/hapi-pino.js
